+++ packages/core/src/styled/mixins/state/elevationBorderMixin.ts
@@ -1,6 +1,7 @@
 import { css, DefaultTheme } from 'styled-components';
 
 import { Elevation } from '../../../declarations';
+import { getTemplate } from './elevationBorderMixinTemplate';
 
 /**
  * Get the border elevation styles based in the theme design tokens and type
@@ -8,30 +9,23 @@ import { Elevation } from '../../../declarations';
 export const elevationBorderMixin =
   (theme: DefaultTheme) => (elevation: Elevation) => {
     const elevations = theme.alias.color.border.elevation;
+    const colorMap = {
+      stickyTop: { color: elevations.sticky.top, side: 'top' },
+      stickyRight: { color: elevations.sticky.right, side: 'right' },
+      stickyLeft: { color: elevations.sticky.left, side: 'left' },
+      stickyBottom: { color: elevations.sticky.bottom, side: 'bottom' },
+    };
     const elevationBorder = theme.alias.shape.borderSize.panel.base;
     return !elevation || elevation === 'ground'
       ? null
       : elevation?.startsWith('sticky')
-        ? elevation === 'stickyTop'
-          ? css`
-              border-top: ${elevationBorder} solid ${elevations.sticky.top};
-            `
-          : elevation === 'stickyRight'
-            ? css`
-                border-right: ${elevationBorder} solid
-                  ${elevations.sticky.right};
-              `
-            : elevation === 'stickyLeft'
-              ? css`
-                  border-left: ${elevationBorder} solid
-                    ${elevations.sticky.left};
-                `
-              : elevation === 'stickyBottom'
-                ? css`
-                    border-bottom: ${elevationBorder} solid
-                      ${elevations.sticky.bottom};
-                  `
-                : null
+        ? Object.keys(colorMap).includes(elevation)
+          ? getTemplate(
+              colorMap[elevation].side,
+              elevationBorder,
+              colorMap[elevation].color,
+            )
+          : null
         : Object.keys(elevations).includes(elevation)
           ? css`
               border: ${elevationBorder} solid ${elevations[elevation]};