+++ packages/core/src/styled/mixins/state/elevationBorderMixinTemplate.ts
@@ -1,0 +1,7 @@
+import { css } from 'styled-components';
+
+export const getTemplate = (
+  side: 'top' | 'left' | 'bottom' | 'right',
+  width: string,
+  color: string,
+) => css`border-${side}: ${width} solid ${color};`;