+++ packages/table/src/helpers/orderDataByOrderStruct/index.ts
@@ -1,2 +1,1 @@
-export * from './getOptionsFromData';
 export * from './orderDataByOrderStruct';