+++ packages/table/src/helpers/orderDataByOrderStruct/sorters/index.ts
@@ -1,0 +1,2 @@
+export * from './numberSorter';
+export * from './stringSorter';