+++ packages/table/src/helpers/orderDataByOrderStruct/sorters/numberSorter.ts
@@ -1,0 +1,5 @@
+export const numberSorter = (
+  a: number,
+  b: number,
+  sort: 'asc' | 'desc' = 'asc',
+) => (sort === 'asc' ? a - b : b - a);